Open shift creation complete api testing left

+++ kairosActivity/src/main/java/com/kairos/activity/persistence/model/priority_group/PriorityGroup.java
@@ -22,7 +22,8 @@ public class PriorityGroup extends MongoBaseEntity {
     private Long unitId;
     private BigInteger countryParentId;
     private Integer priority;
-    private Integer orderId;
+    private BigInteger orderId;
+
 
     public PriorityGroup() {
         //Default Constructor
@@ -113,12 +114,14 @@ public class PriorityGroup extends MongoBaseEntity {
     public void setPriority(Integer priority) {
         this.priority = priority;
     }
-
-    public Integer getOrderId() {
+    public BigInteger getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(BigInteger orderId) {
         this.orderId = orderId;
     }
+
+
+
 }
